requirements.txt
aiohttp==3.11.7
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.20
click==8.1.7
coloredlogs==15.0.1
decorator==5.1.1
docx2txt
frozenlist==1.5.0
gradio==5.6.0
gradio_client==1.4.3
groq==0.12.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
langchain==0.3.7
langchain-chroma==0.1.4
langchain-community==0.3.7
langchain-core==0.3.19
langchain-groq==0.2.1
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langgraph==0.2.53
langgraph-checkpoint==2.0.5
langgraph-sdk==0.1.36
langsmith==0.1.144
multidict==6.1.0
numpy==1.26.4
openai==1.55.0
pydantic==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.27.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.20.3
tomli==2.1.0
tomlkit==0.12.0
typing-inspect==0.9.0
typing_extensions==4.12.2
uvicorn==0.32.1
yarl==1.18.0
streamlit
duckdb